Refactor LLM backend structure and introduce OpenAI integration
- Removed the existing LLM backend implementations (HuggingFace, MLX) and their associated files.
- Introduced a new OpenAI backend for generating text using the OpenAI API, including structured output support.
- Updated the generation logic to utilize the OpenAI backend exclusively, simplifying the configuration and initialization process.
- Enhanced the WritingAgent to default to the OpenAI backend with updated model configurations.
- Adjusted imports and module structure to reflect the removal of deprecated backends and the addition of the OpenAI integration.
